pyproject.toml
[project]
name = "f1bot"
version = "0.1.0"
description = "F1 Telegram Bot - MVP"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "apscheduler>=3.10.0",
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
]

[project.scripts]
f1bot = "f1bot.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["f1bot*"]
